chore: apply clippy fixes and eliminate remaining warnings
- cargo clippy --fix: 90 files, mostly idiomatic improvements
  - `contains` pattern replacements
  - `is_multiple_of` instead of manual `% == 0`
  - `div_ceil` instead of manual implementation
  - Derived `Default` impls where applicable
  - Removed unnecessary parentheses, clones, and closures
  - `if let` simplifications
- Remove unnecessary parentheses in combat.rs exp formula
- Remove unnecessary `mut` binding in animation_system.rs

+++ Client-Macroquad/src/components/combat.rs
@@ -330,7 +330,7 @@ impl Experience {
     
     /// 传奇升级经验公式
     fn calculate_required(level: u16) -> i64 {
-        ((level as i64 + 1) * (level as i64 + 1) * 100) as i64
+        (level as i64 + 1) * (level as i64 + 1) * 100
     }
 }
